Extract the invitation exchange into a shared InviteCompletion collaborator
The exchange - legacy and attested protocols, claims forwarding, recipient
binding, tenant matching and duplicate-subject handling - moves out of
InviteMiddleware into InviteCompletion behind IInviteCompletion, so a second
caller can complete an invitation identically. The middleware keeps its
released constructor contract and phase behavior unchanged, delegating the
exchange and the post-success lobby-redirect resolution to the collaborator.
Tenant resolution falls back to ITenantResolver when the tenancy middleware
has not run for the current request, which answers the same question from the
same request facts.
